Italian translation for WinBox-Reloaded
This is an Italian translation for WinBox-Reloaded. It includes Italian text and messages for all text in WinBox-Reloaded. It is mostly confirmed working, the only thing that doesn't work as of this pull request is the New Virtual Hard Disk Wizard, with it freezing the program when opened. 
This language file can be tested out by changing the extension from ".it-IT" to ".en-US" and replacing the existing "Language.en-US" file within the Languages folder of your WinBox installation.